Cleanup extraneous license notices Update TalkJSClient attributes for nicer interfacing Move event models to Events namespace Base ConversationEvent class for some Conversation events
