+++ src/Enumerations/ConversationPermission.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CarAndClassic\TalkJS\Enumerations;
+
+
+final class ConversationPermission
+{
+    public const READ_WRITE = 'ReadWrite';
+    public const READ = 'read';
+}
