+++ src/Events/ConversationCreatedOrUpdated.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CarAndClassic\TalkJS\Models;
+namespace CarAndClassic\TalkJS\Events;
 
 class ConversationCreatedOrUpdated
 {
