+++ src/Events/ConversationDeleted.php
@@ -2,8 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CarAndClassic\TalkJS\Models;
-
+namespace CarAndClassic\TalkJS\Events;
 
 class ConversationDeleted
 {
