+++ src/Events/ConversationLeft.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CarAndClassic\TalkJS\Events;
+
+class ConversationLeft extends ConversationEvent
+{
+}
