+++ src/Exceptions/ResponseException.php
@@ -2,11 +2,6 @@
 
 declare(strict_types=1);
 
-/*
- * This software may be modified and distributed under the terms
- * of the MIT license. See the LICENSE file for details.
- */
-
 namespace CarAndClassic\TalkJS\Exceptions;
 
 class ResponseException extends \RuntimeException
