--- src/Models/ConversationJoined.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace CarAndClassic\TalkJS\Models;
-
-class ConversationJoined
-{
-}
